refactor(zaino-state): move accumulator tests into the gated module
Relocate the 5 accumulator integration tests + 2 helpers (updates_on_write/delete, the
bulk/incremental builder tests, batched_sync, sync_with_batch_budget,
accumulator_from_unspent_map) out of tests/.../v1.rs into tx_out_set_accumulator.rs's
`#[cfg(test)] mod tests`, beside the unit tests. The module's feature gate covers them, so
the per-test #[cfg(feature = "gettxoutsetinfo")] attributes are dropped. All accumulator
tests now live in the gated module; only the v1.1->v1.2 migration tests keep accumulator
assertions in-place (they also assert spent/txid backfill, which runs when the feature is off).

Widen indexed_block_chain to pub(crate) so the module's tests can reach it. Both feature
states build and pass clippy for lib and tests.
