+++ packages/zaino-state/src/chain_index/finalised_state/finalised_source/v1/tx_out_set_accumulator.rs
@@ -1306,9 +1306,21 @@ impl DbV1 {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::chain_index::finalised_state::FinalisedState;
+    use crate::chain_index::tests::finalised_state::v1::load_vectors_and_spawn_and_sync_v1_zaino_db;
+    use crate::chain_index::tests::init_tracing;
+    use crate::chain_index::tests::vectors::{
+        build_mockchain_source, indexed_block_chain, load_test_vectors, TestVectorBlockData,
+        TestVectorData,
+    };
     use crate::chain_index::types::db::metadata::{
         FinalisedTxOutSetInfoAccumulator, ZAINO_TXOUTSET_ENTRY_LEN,
     };
+    use crate::ChainIndexConfig;
+    use std::sync::Arc;
+    use tempfile::TempDir;
+    use zaino_common::network::ActivationHeights;
+    use zaino_common::{DatabaseConfig, Network, StorageConfig};
 
     fn p2pkh_out(value: u64) -> TxOutCompact {
         TxOutCompact::new(value, [0x11; 20], 0).expect("P2PKH script_type should be valid")
@@ -1489,5 +1501,450 @@ mod tests {
         .expect("reverse should succeed");
 
         assert_eq!(acc, FinalisedTxOutSetInfoAccumulator::empty());
+    }
+
+    #[tokio::test(flavor = "multi_thread")]
+    async fn tx_out_set_info_accumulator_updates_on_write() {
+        init_tracing();
+
+        // Load the regtest vectors, write every vector block into FinalisedState, and wait until the
+        // finalised state has finished its startup/background validation.
+        let (TestVectorData { blocks, .. }, _db_dir, zaino_db) =
+            load_vectors_and_spawn_and_sync_v1_zaino_db().await;
+
+        zaino_db.wait_until_ready().await;
+
+        let db_reader = Arc::new(zaino_db).to_reader();
+
+        // Build the expected UTXO set directly from the same vector blocks.
+        //
+        // Map shape:
+        //   txid -> { output_index -> TxOutCompact }
+        //
+        // From this we derive every accumulator field:
+        //   transactions         = number of txids with at least one unspent output
+        //   transaction_outputs  = total number of unspent transparent outputs
+        //   bytes_serialized     = transaction_outputs * ZAINO_TXOUTSET_ENTRY_LEN
+        //   hash_serialized      = XOR of tx_out_set_entry_digest over all unspent outputs
+        //   total_zatoshis       = sum of `value` over all unspent outputs
+        let mut unspent_output_indices_by_transaction_hash: HashMap<
+            TransactionHash,
+            HashMap<u32, crate::TxOutCompact>,
+        > = HashMap::new();
+
+        for chain_block in indexed_block_chain(&blocks) {
+            for transaction in chain_block.transactions() {
+                // First apply spends, removing spent transparent outputs from the expected UTXO set.
+                for input in transaction.transparent().inputs() {
+                    if input.is_null_prevout() {
+                        continue;
+                    }
+
+                    let previous_transaction_hash = TransactionHash::from(*input.prevout_txid());
+
+                    let unspent_output_indices = unspent_output_indices_by_transaction_hash
+                        .get_mut(&previous_transaction_hash)
+                        .unwrap_or_else(|| {
+                            panic!(
+                            "test vectors spend unknown transaction {previous_transaction_hash:?}"
+                        )
+                        });
+
+                    assert!(
+                        unspent_output_indices
+                            .remove(&input.prevout_index())
+                            .is_some(),
+                        "test vectors spend unknown output: transaction {:?}, output {}",
+                        previous_transaction_hash,
+                        input.prevout_index()
+                    );
+
+                    // If a transaction has no remaining unspent outputs, it should no longer
+                    // contribute to the accumulator's `transactions` count.
+                    if unspent_output_indices.is_empty() {
+                        unspent_output_indices_by_transaction_hash
+                            .remove(&previous_transaction_hash);
+                    }
+                }
+
+                // Then apply outputs, adding newly-created transparent outputs to the expected UTXO set.
+                if transaction.transparent().outputs().is_empty() {
+                    continue;
+                }
+
+                let transaction_hash = *transaction.txid();
+
+                let unspent_output_indices = unspent_output_indices_by_transaction_hash
+                    .entry(transaction_hash)
+                    .or_default();
+
+                for (output_index, output) in transaction.transparent().outputs().iter().enumerate()
+                {
+                    // The accumulator skips NonStandard (unspendable) outputs — see
+                    // `is_unspendable_tx_out` in
+                    // `chain_index::types::db::metadata`. The oracle must mirror that.
+                    if crate::chain_index::types::db::metadata::is_unspendable_tx_out(output) {
+                        continue;
+                    }
+
+                    let output_index = u32::try_from(output_index).unwrap();
+
+                    assert!(
+                    unspent_output_indices
+                        .insert(output_index, *output)
+                        .is_none(),
+                    "test vectors duplicate output index: transaction {transaction_hash:?}, output {output_index}"
+                );
+                }
+
+                // If the transaction had only NonStandard outputs, drop the empty entry so it
+                // doesn't inflate the expected `transactions` count.
+                if unspent_output_indices.is_empty() {
+                    unspent_output_indices_by_transaction_hash.remove(&transaction_hash);
+                }
+            }
+        }
+
+        let expected_accumulator =
+            accumulator_from_unspent_map(&unspent_output_indices_by_transaction_hash);
+
+        // Check that the accumulator maintained by write_block matches the independently
+        // reconstructed expected UTXO-set counts.
+        let actual_accumulator = db_reader.get_tx_out_set_info_accumulator().await.unwrap();
+
+        assert_eq!(expected_accumulator, actual_accumulator);
+    }
+
+    /// The bulk sequential accumulator builder must produce exactly the accumulator that the
+    /// per-block incremental write path maintained, for every shard count. Sharding partitions the
+    /// work by creating-txid prefix and recombines the partials; the result must be shard-count
+    /// independent.
+    #[tokio::test(flavor = "multi_thread")]
+    async fn bulk_tx_out_set_accumulator_builder_matches_incremental() {
+        init_tracing();
+
+        let (_data, _db_dir, zaino_db) = load_vectors_and_spawn_and_sync_v1_zaino_db().await;
+        zaino_db.wait_until_ready().await;
+
+        use crate::chain_index::finalised_state::capability::{
+            CapabilityRequest, DbRead, TransparentHistExt,
+        };
+
+        let backend = zaino_db
+            .backend_for_cap(CapabilityRequest::WriteCore)
+            .unwrap();
+
+        let db_tip = backend.db_height().await.unwrap().unwrap();
+        let incremental = backend.get_tx_out_set_info_accumulator().await.unwrap();
+
+        // 1 = single optimal pass; >1 exercises the sharded multi-pass recombination; 256 = one
+        // first-byte value per shard (maximal sharding).
+        for shards in [1u16, 2, 4, 256] {
+            let built = tokio::task::block_in_place(|| {
+                backend.build_tx_out_set_accumulator_blocking(db_tip, shards)
+            })
+            .unwrap();
+
+            assert_eq!(
+            built, incremental,
+            "bulk builder (shards={shards}) must equal the incrementally-maintained accumulator"
+        );
+        }
+    }
+
+    /// Syncs the vector chain to height 200 with the given bulk-write batch budget and returns the
+    /// resulting `(db tip, validated tip, txout-set accumulator)`.
+    async fn sync_with_batch_budget(
+        blocks: Vec<TestVectorBlockData>,
+        sync_write_batch_bytes: u64,
+    ) -> (Height, u32, FinalisedTxOutSetInfoAccumulator) {
+        use crate::chain_index::finalised_state::capability::{
+            CapabilityRequest, DbRead, TransparentHistExt,
+        };
+
+        let source = build_mockchain_source(blocks);
+        let temp_dir: TempDir = tempfile::tempdir().unwrap();
+        let config = ChainIndexConfig {
+            storage: StorageConfig {
+                database: DatabaseConfig {
+                    path: temp_dir.path().to_path_buf(),
+                    sync_write_batch_bytes,
+                    ..Default::default()
+                },
+                ..Default::default()
+            },
+            ephemeral: false,
+            db_version: 1,
+            network: Network::Regtest(ActivationHeights::default()),
+        };
+
+        let zaino_db = FinalisedState::spawn(config, source.clone()).await.unwrap();
+        zaino_db.sync_to_height(Height(200), &source).await.unwrap();
+        // Catch-up of >LONG_RUNNING_SYNC_THRESHOLD blocks runs in the background; wait for the
+        // persistent DB to actually reach the tip before reading it back.
+        zaino_db.wait_until_synced().await;
+
+        let backend = zaino_db
+            .backend_for_cap(CapabilityRequest::WriteCore)
+            .unwrap();
+        let db_tip = backend.db_height().await.unwrap().unwrap();
+        let validated_tip = backend.validated_tip_height();
+        let accumulator = backend.get_tx_out_set_info_accumulator().await.unwrap();
+
+        (db_tip, validated_tip, accumulator)
+    }
+
+    /// The bulk-sync result must be independent of the write-batch budget: a single huge batch and a
+    /// one-block-per-batch sync of the same chain must produce an identical db tip, validated tip, and
+    /// txout-set accumulator. This exercises the cross-batch continuity chaining, per-batch
+    /// `validated_tip` advance, and sorted-insert flush boundaries that a single-batch sync does not.
+    #[tokio::test(flavor = "multi_thread")]
+    async fn batched_sync_is_batch_size_independent() {
+        init_tracing();
+
+        let blocks = load_test_vectors().unwrap().blocks;
+
+        // u64::MAX => the whole sync is one batch; 1 => every block exceeds the budget => one block per
+        // batch (a flush + commit + fsync after each block).
+        let single_batch = sync_with_batch_budget(blocks.clone(), u64::MAX).await;
+        let per_block_batches = sync_with_batch_budget(blocks, 1).await;
+
+        assert_eq!(single_batch.0, per_block_batches.0, "db tip must match");
+        assert_eq!(
+            single_batch.1, per_block_batches.1,
+            "validated tip must match"
+        );
+        assert_eq!(
+            single_batch.2, per_block_batches.2,
+            "txout-set accumulator must be independent of the write-batch budget"
+        );
+    }
+
+    /// The incremental range-update path — taken when a catch-up advances an already-built accumulator
+    /// by a small range (`write_blocks_to_height`'s steady-state branch) — must produce exactly the
+    /// accumulator a full from-genesis rebuild produces at the same tip, for all five fields. This is
+    /// the correctness gate for `update_tx_out_set_accumulator_for_range`: with regtest coinbase
+    /// maturity of 100, splitting the sync at height 100 guarantees the second segment spends outputs
+    /// created in the first (exercising the `transactions` "Set B" decrement) as well as outputs both
+    /// created and spent within the range (the XOR-cancel case).
+    #[tokio::test(flavor = "multi_thread")]
+    async fn incremental_accumulator_update_matches_full_rebuild() {
+        init_tracing();
+
+        use crate::chain_index::finalised_state::capability::{
+            CapabilityRequest, DbRead, TransparentHistExt,
+        };
+
+        let blocks = load_test_vectors().unwrap().blocks;
+        let source = build_mockchain_source(blocks);
+        let temp_dir: TempDir = tempfile::tempdir().unwrap();
+        let config = ChainIndexConfig {
+            storage: StorageConfig {
+                database: DatabaseConfig {
+                    path: temp_dir.path().to_path_buf(),
+                    ..Default::default()
+                },
+                ..Default::default()
+            },
+            ephemeral: false,
+            db_version: 1,
+            network: Network::Regtest(ActivationHeights::default()),
+        };
+
+        let zaino_db = FinalisedState::spawn(config, source.clone()).await.unwrap();
+
+        // First segment builds the accumulator to height 100 (no watermark yet => full rebuild),
+        // the second advances it by a 100-block range => the incremental update path under test.
+        zaino_db.sync_to_height(Height(100), &source).await.unwrap();
+        // Background catch-up (>LONG_RUNNING_SYNC_THRESHOLD); wait for the persistent build + watermark.
+        zaino_db.wait_until_synced().await;
+
+        let backend = zaino_db
+            .backend_for_cap(CapabilityRequest::WriteCore)
+            .unwrap();
+
+        // The watermark must sit at 100 here: that (together with gap 100 <= the incremental cap)
+        // pins the next sync to the incremental branch rather than a silent rebuild fallback that
+        // would make the comparison below trivial.
+        assert_eq!(
+            backend
+                .read_tx_out_set_accumulator_built_height()
+                .await
+                .unwrap(),
+            Some(Height(100)),
+            "first segment must leave the accumulator watermark at the synced tip"
+        );
+
+        zaino_db.sync_to_height(Height(200), &source).await.unwrap();
+        // Background catch-up; wait for the incremental accumulator update to advance the watermark.
+        zaino_db.wait_until_synced().await;
+
+        let db_tip = backend.db_height().await.unwrap().unwrap();
+        assert_eq!(db_tip, Height(200), "both segments must have been synced");
+        assert_eq!(
+            backend
+                .read_tx_out_set_accumulator_built_height()
+                .await
+                .unwrap(),
+            Some(Height(200)),
+            "incremental update must advance the watermark to the new tip"
+        );
+
+        let incremental = backend.get_tx_out_set_info_accumulator().await.unwrap();
+        let from_genesis = tokio::task::block_in_place(|| {
+            backend.build_tx_out_set_accumulator_blocking(db_tip, 1)
+        })
+        .unwrap();
+
+        assert_eq!(
+            incremental, from_genesis,
+            "incremental range-update accumulator must equal the from-genesis rebuild at the tip"
+        );
+    }
+
+    /// Computes the canonical [`FinalisedTxOutSetInfoAccumulator`] for a fully-resolved UTXO set,
+    /// used as the source of truth by the write/delete accumulator tests.
+    fn accumulator_from_unspent_map(
+        unspent: &HashMap<TransactionHash, HashMap<u32, crate::TxOutCompact>>,
+    ) -> FinalisedTxOutSetInfoAccumulator {
+        use crate::chain_index::types::db::metadata::{
+            tx_out_set_entry_digest, ZAINO_TXOUTSET_ENTRY_LEN,
+        };
+        use crate::Outpoint;
+
+        let mut transaction_outputs = 0u64;
+        let mut total_zatoshis = 0u64;
+        let mut hash_serialized = [0u8; 32];
+
+        for (txid, outputs) in unspent {
+            for (output_index, out) in outputs {
+                let outpoint = Outpoint::new(txid.0, *output_index);
+                let digest = tx_out_set_entry_digest(&outpoint, out);
+                for (dst, src) in hash_serialized.iter_mut().zip(digest.iter()) {
+                    *dst ^= *src;
+                }
+                transaction_outputs += 1;
+                total_zatoshis += out.value();
+            }
+        }
+
+        FinalisedTxOutSetInfoAccumulator {
+            transactions: unspent.len() as u64,
+            transaction_outputs,
+            bytes_serialized: transaction_outputs * ZAINO_TXOUTSET_ENTRY_LEN,
+            hash_serialized,
+            total_zatoshis,
+        }
+    }
+
+    #[tokio::test(flavor = "multi_thread")]
+    async fn tx_out_set_info_accumulator_updates_on_delete() {
+        init_tracing();
+
+        // Load and write all vector blocks, then delete the current tip block.
+        // The accumulator should end up matching the UTXO set for all blocks except the deleted tip.
+        let (TestVectorData { blocks, .. }, _db_dir, zaino_db) =
+            load_vectors_and_spawn_and_sync_v1_zaino_db().await;
+
+        zaino_db.wait_until_ready().await;
+
+        let deleted_block_height = Height(blocks.last().unwrap().height);
+
+        zaino_db
+            .delete_block_at_height(deleted_block_height)
+            .await
+            .unwrap();
+
+        zaino_db.wait_until_ready().await;
+
+        let db_reader = Arc::new(zaino_db).to_reader();
+
+        // Rebuild the expected UTXO set from the vector chain with the deleted tip excluded.
+        //
+        // This verifies that delete_block reverses every accumulator field:
+        //   transactions, transaction_outputs, bytes_serialized, hash_serialized, total_zatoshis.
+        let mut unspent_output_indices_by_transaction_hash: HashMap<
+            TransactionHash,
+            HashMap<u32, crate::TxOutCompact>,
+        > = HashMap::new();
+
+        for chain_block in indexed_block_chain(&blocks[..blocks.len() - 1]) {
+            for transaction in chain_block.transactions() {
+                // Remove any transparent outputs spent by this transaction.
+                for input in transaction.transparent().inputs() {
+                    if input.is_null_prevout() {
+                        continue;
+                    }
+
+                    let previous_transaction_hash = TransactionHash::from(*input.prevout_txid());
+
+                    let unspent_output_indices = unspent_output_indices_by_transaction_hash
+                        .get_mut(&previous_transaction_hash)
+                        .unwrap_or_else(|| {
+                            panic!(
+                            "test vectors spend unknown transaction {previous_transaction_hash:?}"
+                        )
+                        });
+
+                    assert!(
+                        unspent_output_indices
+                            .remove(&input.prevout_index())
+                            .is_some(),
+                        "test vectors spend unknown output: transaction {:?}, output {}",
+                        previous_transaction_hash,
+                        input.prevout_index()
+                    );
+
+                    if unspent_output_indices.is_empty() {
+                        unspent_output_indices_by_transaction_hash
+                            .remove(&previous_transaction_hash);
+                    }
+                }
+
+                // Add this transaction's newly-created transparent outputs.
+                if transaction.transparent().outputs().is_empty() {
+                    continue;
+                }
+
+                let transaction_hash = *transaction.txid();
+
+                let unspent_output_indices = unspent_output_indices_by_transaction_hash
+                    .entry(transaction_hash)
+                    .or_default();
+
+                for (output_index, output) in transaction.transparent().outputs().iter().enumerate()
+                {
+                    // The accumulator skips NonStandard (unspendable) outputs — see
+                    // `is_unspendable_tx_out` in
+                    // `chain_index::types::db::metadata`. The oracle must mirror that.
+                    if crate::chain_index::types::db::metadata::is_unspendable_tx_out(output) {
+                        continue;
+                    }
+
+                    let output_index = u32::try_from(output_index).unwrap();
+
+                    assert!(
+                    unspent_output_indices
+                        .insert(output_index, *output)
+                        .is_none(),
+                    "test vectors duplicate output index: transaction {transaction_hash:?}, output {output_index}"
+                );
+                }
+
+                // If the transaction had only NonStandard outputs, drop the empty entry so it
+                // doesn't inflate the expected `transactions` count.
+                if unspent_output_indices.is_empty() {
+                    unspent_output_indices_by_transaction_hash.remove(&transaction_hash);
+                }
+            }
+        }
+
+        let expected_accumulator =
+            accumulator_from_unspent_map(&unspent_output_indices_by_transaction_hash);
+
+        // Check the accumulator persisted by delete_block_at_height/delete_block.
+        let actual_accumulator = db_reader.get_tx_out_set_info_accumulator().await.unwrap();
+
+        assert_eq!(expected_accumulator, actual_accumulator);
     }
 }
